really fix docker
