+++ Dockerfile
@@ -4,17 +4,22 @@
 
 FROM public.ecr.aws/docker/library/python:3.13-alpine AS builder
 
-RUN apk add build-base linux-headers 
+# NOTE1: build-base and linux-headers are required to build psutil dependency
+# NOTE2: git is required to get the "hatch-vcs" plugin to work and produce the _psmqtt_version.py file
+RUN apk add build-base linux-headers git
 
 WORKDIR /build
-COPY ./requirements.txt .
-#COPY ./pyproject.toml .
-#COPY ./src .
-#COPY ./README.rst .
+COPY requirements.txt pyproject.toml README.rst ./
+COPY ./src ./src/
+COPY ./.git ./.git/
 
 RUN python -m pip install --upgrade pip
 RUN pip install --target=/build/deps -r requirements.txt
-#RUN rm -r requirements.txt
+RUN pip install build
+#RUN git config --global --add safe.directory /build
+#RUN mkdir src
+RUN python -m build --wheel --outdir /build/wheel
+#RUN hatch build
 
 
 #
@@ -29,14 +34,19 @@ ARG USERNAME=psmqtt
 
 LABEL org.opencontainers.image.source=https://github.com/f18m/psmqtt
 
+# install pySMART (and bash to ease debug -- TODO get rid of bash)
 RUN apk add bash smartmontools
 
 WORKDIR /opt/psmqtt
-COPY --from=builder /build .
 
-RUN mkdir ./src
-COPY *.py ./
-COPY src/psmqtt/*.py ./src
+# copy the dependencies from the builder stage
+COPY --from=builder /build/deps ./deps/
+
+# copy all source code
+COPY src/psmqtt/*.py ./src/
+
+# copy the version file produced by hatch-vcs plugin from the builder stage:
+COPY --from=builder /build/src/_psmqtt_version.py ./src/
 
 RUN mkdir ./conf ./schema
 
@@ -63,7 +73,10 @@ ENV PSMQTTCONFIG="/opt/psmqtt/conf/psmqtt.yaml"
 ENV PSMQTTCONFIGSCHEMA="/opt/psmqtt/schema/psmqtt.schema.yaml"
 
 # add deps to PYTHONPATH
-ENV PYTHONPATH="/opt/psmqtt/deps"
+ENV PYTHONPATH="/opt/psmqtt/src:/opt/psmqtt/deps"
 
 # run process
-CMD ["python", "psmqtt.py"]
+# it's important to use python -m to run the module, otherwise the relative imports
+# will not work. Remember that the docker image does not contain the actual psmqtt
+# wheel installed (this is to make it possible to remove "pip" from the base image in future)
+CMD ["python", "-m", "src.main"]
